+++ app/Http/Controllers/ProjectController.php
@@ -9,12 +9,15 @@ use App\Models\Tilsyn_object;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class ProjectController extends Controller
 {
     public function index()
     {
+        Gate::authorize('project_show');
+
         $projects = Project::with('tilsyn_objects')->orderBy('id', 'DESC')->get();
         $noProjects = Tilsyn_object::where('project_id', null)->get();
 
@@ -26,6 +29,8 @@ class ProjectController extends Controller
 
     public function store(StoreProjectRequest $request): RedirectResponse
     {
+        Gate::authorize('project_edit');
+
         Project::create($request->validated());
 
         return to_route('projects')
@@ -34,6 +39,8 @@ class ProjectController extends Controller
 
     public function destroy(Project $project): RedirectResponse
     {
+        Gate::authorize('project_edit');
+
         $project->delete();
 
         Tilsyn_object::where('project_id', $project->id)->update(['project_id' => null]);
